+++ lib/example/SimpleApp.d.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from "express";
 import { ApplicationInstance } from "../index";
+import { HttpContext } from "../easy-rest/http/http-context";
 export declare class SimpleApp extends ApplicationInstance {
     constructor();
+    onError(error: any, httpContext: HttpContext): void;
     simpleHandler(req: Request, res: Response): Promise<void>;
-    simpleErrorHandler1(err: any, req: Request, res: Response): Promise<any>;
-    simpleErrorHandler2(err: any, req: Request, res: Response): Promise<any>;
     private getAuthProvider();
 }